backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "dotenv>=0.9.9",
    "duckduckgo-search>=8.0.4",
    "fastapi>=0.115.12",
    "google-search-results>=2.4.2",
    "groq>=0.28.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-groq>=0.3.2",
    "langgraph>=0.4.8",
    "logging>=0.4.9.6",
    "newspaper3k>=0.2.8",
    "nltk>=3.9.1",
    "pinecone>=7.3.0",
    "rake-nltk>=1.0.6",
    "readability-lxml>=0.8.4.1",
    "requests>=2.32.3",
    "sentence-transformers>=5.0.0",
    "trafilatura>=2.0.0",
    "uvicorn>=0.34.3",
    "langchain-openai>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "pydantic-settings>=2.0.0",
]
